% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Poly}
\alias{Poly}
\title{Polynomial generation and multiplication}
\usage{
Poly(s)
}
\arguments{
\item{s}{a vector specifying the the coefficients of the polynomial}
}
\value{
An object of class \code{Poly} giving the coefficients of the polynomial.
}
\description{
Generate a polynomial object.
}
\details{
There are 2 methods for multiplication and one for printing
}
\examples{
## Generate the sequence of the first 10 Fibonacci numbers.
p <- Poly(c(1,2,3))
## One can now print:
print(p)
}
\references{
\url{https://en.wikipedia.org/wiki/Polynomial}.
}
\author{
Jakub Cery and Andrea Srnakova
}
